feat: authorization in all routes

+++ src/album/album.controller.ts
@@ -9,23 +9,39 @@ import {
   HttpException,
   HttpCode,
   Res,
+  Req,
   HttpStatus,
 } from '@nestjs/common';
-import { Response } from 'express';
+import { Request, Response } from 'express';
 import { AlbumService } from './album.service';
 import { Album } from 'src/common/types';
+import { accessCheck } from 'src/common/access';
 
 @Controller('album')
 export class AlbumController {
   constructor(private albumService: AlbumService) {}
 
   @Get()
-  async getAllAlbums(@Res() res: Response): Promise<string> {
+  async getAllAlbums(
+    @Res() res: Response,
+    @Req() request: Request,
+  ): Promise<string> {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     res.status(HttpStatus.OK).send(await this.albumService.getAllAlbums());
     return '';
   }
   @Get(':id')
-  async getAlbumById(@Param('id') id: string): Promise<string> {
+  async getAlbumById(
+    @Param('id') id: string,
+    @Req() request: Request,
+  ): Promise<string> {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const dbResponse = await this.albumService.getAlbum(id);
     const data = dbResponse.data as Album;
     if (!data || !data.id) {
@@ -34,7 +50,15 @@ export class AlbumController {
     return JSON.stringify(dbResponse.data);
   }
   @Post()
-  async create(@Body() album: Album, @Res() res: Response) {
+  async create(
+    @Body() album: Album,
+    @Res() res: Response,
+    @Req() request: Request,
+  ) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const result = await this.albumService.create(album);
     if (result.code === 400)
       throw new HttpException('Data missing', result.code);
@@ -45,7 +69,12 @@ export class AlbumController {
     @Param('id') id: string,
     @Body() album: Album,
     @Res() res: Response,
+    @Req() request: Request,
   ) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const result = await this.albumService.changeAlbum(id, album);
     if (result.code !== 200)
       throw new HttpException(result.message, result.code);
@@ -53,7 +82,11 @@ export class AlbumController {
   }
   @Delete(':id')
   @HttpCode(204)
-  async delUserById(@Param('id') id: string) {
+  async delUserById(@Param('id') id: string, @Req() request: Request) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const result = await this.albumService.removeAlbum(id);
     if (result.code !== 200)
       throw new HttpException(result.message, result.code);
