+++ src/favs/favs.controller.ts
@@ -7,21 +7,38 @@ import {
   HttpException,
   HttpCode,
   Res,
+  Req,
   HttpStatus,
 } from '@nestjs/common';
-import { Response } from 'express';
+import { Response, Request } from 'express';
 import { FavService } from './favor.service';
+import { accessCheck } from 'src/common/access';
 
 @Controller('favs')
 export class FavsController {
   constructor(private favService: FavService) {}
   @Get()
-  async getAllFavs(@Res() res: Response): Promise<string> {
+  async getAllFavs(
+    @Res() res: Response,
+    @Req() request: Request,
+  ): Promise<string> {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     res.status(HttpStatus.OK).send(await this.favService.getAllFavs());
     return '';
   }
   @Post('track/:id')
-  async addTrack(@Param('id') id: string, @Res() res: Response) {
+  async addTrack(
+    @Param('id') id: string,
+    @Res() res: Response,
+    @Req() request: Request,
+  ) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const result = await this.favService.addTrack(id);
     if (result.code === 400)
       throw new HttpException('Data missing', result.code);
@@ -30,7 +47,15 @@ export class FavsController {
     res.status(HttpStatus.CREATED).send(result.data);
   }
   @Post('album/:id')
-  async addAlbum(@Param('id') id: string, @Res() res: Response) {
+  async addAlbum(
+    @Param('id') id: string,
+    @Res() res: Response,
+    @Req() request: Request,
+  ) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const result = await this.favService.addAlbum(id);
     if (result.code === 400)
       throw new HttpException('Data missing', result.code);
@@ -39,7 +64,15 @@ export class FavsController {
     res.status(HttpStatus.CREATED).send(result.data);
   }
   @Post('artist/:id')
-  async addArtist(@Param('id') id: string, @Res() res: Response) {
+  async addArtist(
+    @Param('id') id: string,
+    @Res() res: Response,
+    @Req() request: Request,
+  ) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const result = await this.favService.addArtist(id);
     if (result.code === 400)
       throw new HttpException('Data missing', result.code);
@@ -50,7 +83,11 @@ export class FavsController {
 
   @Delete('track/:id')
   @HttpCode(204)
-  async removeTrack(@Param('id') id: string) {
+  async removeTrack(@Param('id') id: string, @Req() request: Request) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const result = await this.favService.removeTrack(id);
     if (result.code !== 204)
       throw new HttpException(result.message, result.code);
@@ -58,7 +95,11 @@ export class FavsController {
 
   @Delete('album/:id')
   @HttpCode(204)
-  async removeAlbum(@Param('id') id: string) {
+  async removeAlbum(@Param('id') id: string, @Req() request: Request) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const result = await this.favService.removeAlbum(id);
     if (result.code !== 204)
       throw new HttpException(result.message, result.code);
@@ -66,7 +107,11 @@ export class FavsController {
 
   @Delete('artist/:id')
   @HttpCode(204)
-  async removeArtist(@Param('id') id: string) {
+  async removeArtist(@Param('id') id: string, @Req() request: Request) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const result = await this.favService.removeArtist(id);
     if (result.code !== 204)
       throw new HttpException(result.message, result.code);
