+++ src/user/user.controller.ts
@@ -9,24 +9,40 @@ import {
   HttpException,
   HttpCode,
   Res,
+  Req,
   HttpStatus,
 } from '@nestjs/common';
-import { Response } from 'express';
+import { Request, Response } from 'express';
 import { ClearUser } from 'src/common/types';
 import { CreateUserDto, UpdatePasswordDto } from '../dto/user.dto';
 import { UserService } from './user.service';
+import { accessCheck } from 'src/common/access';
 
 @Controller('user')
 export class UserController {
   constructor(private userService: UserService) {}
 
   @Get()
-  async getAllUsers(@Res() res: Response): Promise<string> {
+  async getAllUsers(
+    @Res() res: Response,
+    @Req() request: Request,
+  ): Promise<string> {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     res.status(HttpStatus.OK).send(await this.userService.getAllUsers());
     return '';
   }
   @Get(':id')
-  async getUserById(@Param('id') id: string): Promise<string> {
+  async getUserById(
+    @Param('id') id: string,
+    @Req() request: Request,
+  ): Promise<string> {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const dbResponse = await this.userService.getUser(id);
     const resData = dbResponse.data as ClearUser;
     if (!resData || !resData.id) {
@@ -35,7 +51,15 @@ export class UserController {
     return JSON.stringify(dbResponse.data);
   }
   @Post()
-  async create(@Body() createUserDto: CreateUserDto, @Res() res: Response) {
+  async create(
+    @Body() createUserDto: CreateUserDto,
+    @Req() request: Request,
+    @Res() res: Response,
+  ) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const dbResponse = await this.userService.create(createUserDto);
     if (dbResponse.code === 400)
       throw new HttpException('Data missing', dbResponse.code);
@@ -46,7 +70,12 @@ export class UserController {
     @Param('id') id: string,
     @Body() updatePasswordDto: UpdatePasswordDto,
     @Res() res: Response,
+    @Req() request: Request,
   ) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const result = await this.userService.changePassword(id, updatePasswordDto);
     if (result.code !== 200)
       throw new HttpException(result.message, result.code);
@@ -54,7 +83,11 @@ export class UserController {
   }
   @Delete(':id')
   @HttpCode(204)
-  async delUserById(@Param('id') id: string) {
+  async delUserById(@Param('id') id: string, @Req() request: Request) {
+    const auth = accessCheck(request.headers.authorization);
+    if (auth.code === 401) {
+      throw new HttpException('Not Authorized', auth.code);
+    }
     const result = await this.userService.removeUser(id);
     if (result.code !== 200)
       throw new HttpException(result.message, result.code);
